+++ ivy/bipart.py
@@ -1,7 +1,3 @@
-import sys
-from pprint import pprint
-from glob import glob
-from storage import Storage
 from collections import defaultdict
 
 ## class BipartSet(object):
